move some screening utils out of RateCollection they don't really depend on anything in the RateCollection class
