+++ pynucastro/networks/rate_collection.py
@@ -29,7 +29,7 @@ from pynucastro.rates import (ApproximateRate, DerivedRate, Library, Rate,
                               find_duplicate_rates, is_allowed_dupe, load_rate)
 from pynucastro.rates.library import _rate_name_to_nuc, capitalize_rid
 from pynucastro.screening import make_plasma_state, make_screen_factors
-from pynucastro.screening.screen import NseState
+from pynucastro.screening.screen import NseState, get_screening_map
 
 mpl.rcParams['figure.dpi'] = 100
 
@@ -366,38 +366,6 @@ class Composition:
 
         plt.show()
         return fig
-
-
-class ScreeningPair:
-    """a pair of nuclei that will have rate screening applied.  We store a
-    list of all rates that match this pair of nuclei"""
-
-    def __init__(self, name, nuc1, nuc2, rate=None):
-        self.name = name
-        self.n1 = nuc1
-        self.n2 = nuc2
-
-        if rate is None:
-            self.rates = []
-        else:
-            self.rates = [rate]
-
-    def add_rate(self, rate):
-        if rate not in self.rates:
-            self.rates.append(rate)
-
-    def __str__(self):
-        ostr = f"screening for {self.n1} + {self.n2}\n"
-        ostr += "rates:\n"
-        for r in self.rates:
-            ostr += f"  {r}\n"
-        return ostr
-
-    def __eq__(self, other):
-        """all we care about is whether the names are the same -- that conveys
-        what the reaction is"""
-
-        return self.name == other.name
 
 
 class RateCollection:
@@ -1079,7 +1047,10 @@ class RateCollection:
         factors = {}
         ys = composition.get_molar()
         plasma_state = make_plasma_state(T, rho, ys)
-        screening_map = self.get_screening_map()
+        if not self.do_screening:
+            screening_map = []
+        else:
+            screening_map = get_screening_map(self.get_rates())
 
         for i, scr in enumerate(screening_map):
             if not (scr.n1.dummy or scr.n2.dummy):
@@ -1650,82 +1621,6 @@ class RateCollection:
             ostr += "\n"
 
         return ostr
-
-    def get_screening_map(self):
-        """a screening map is just a list of tuples containing the information
-        about nuclei pairs for screening: (descriptive name of nuclei,
-        nucleus 1, nucleus 2, rate, 1-based index of rate).  If symmetric_screening=True,
-        then for reverse rates, we screen using the forward rate nuclei (assuming that we
-        got here via detailed balance).
-
-        """
-        screening_map = []
-        if not self.do_screening:
-            return screening_map
-
-        # we need to consider the child rates that come with ApproximateRate
-        all_rates = []
-        for r in self.rates:
-            if isinstance(r, ApproximateRate):
-                all_rates += r.get_child_rates()
-            else:
-                all_rates.append(r)
-
-        for r in all_rates:
-            screen_nuclei = r.ion_screen
-            if self.symmetric_screening:
-                screen_nuclei = r.symmetric_screen
-
-            # screen_nuclei may be [] if it is a decay, gamma-capture, or neutron-capture
-            if not screen_nuclei:
-                continue
-
-            nucs = "_".join([str(q) for q in screen_nuclei])
-
-            scr = [q for q in screening_map if q.name == nucs]
-
-            assert len(scr) <= 1
-
-            if scr:
-                # we already have the reactants in our map, so we
-                # will already be doing the screening factors.
-                # Just append this new rate to the list we are
-                # keeping of the rates where this screening is
-                # needed -- if the rate is already in the list, then
-                # this is a no-op
-
-                scr[0].add_rate(r)
-
-                # if we got here because nuc == "He4_He4_He4",
-                # then we also have to add to "He4_He4_He4_dummy"
-
-                if nucs == "He4_He4_He4":
-                    scr2 = [q for q in screening_map if q.name == nucs + "_dummy"]
-                    assert len(scr2) == 1
-
-                    scr2[0].add_rate(r)
-
-            else:
-
-                # we handle 3-alpha specially -- we actually need
-                # 2 screening factors for it
-
-                if nucs == "He4_He4_He4":
-                    # he4 + he4
-                    scr1 = ScreeningPair(nucs, screen_nuclei[0], screen_nuclei[1], r)
-
-                    # he4 + be8
-                    be8 = Nucleus("Be8", dummy=True)
-                    scr2 = ScreeningPair(nucs + "_dummy", screen_nuclei[2], be8, r)
-
-                    screening_map.append(scr1)
-                    screening_map.append(scr2)
-
-                else:
-                    scr1 = ScreeningPair(nucs, screen_nuclei[0], screen_nuclei[1], r)
-                    screening_map.append(scr1)
-
-        return screening_map
 
     def write_network(self, *args, **kwargs):
         """Before writing the network, check to make sure the rates
