+++ pynucastro/screening/__init__.py
@@ -1,6 +1,7 @@
 """Screening routines"""
 
-__all__ = ["screen"]
+__all__ = ["screen", "screening_util"]
 
 from .screen import (PlasmaState, ScreenFactors, chugunov_2007, chugunov_2009,
                      make_plasma_state, make_screen_factors, potekhin_1998)
+from .screening_util import ScreeningPair, get_screening_map
